Move placeholder pages to app folder
